app/requirements.txt
Flask==1.0.2
Flask-SQLAlchemy==2.3.2
Flask-Migrate==2.4.0
Flask-WTF==0.14.2
